move plotting script in specific folder
