+++ benchmarking/nursery/benchmark_kdd/benchmark_nworkers.py
@@ -6,13 +6,11 @@ import logging
 from argparse import ArgumentParser
 from tqdm import tqdm
 
-from benchmarking.blackbox_repository import load
 from benchmarking.blackbox_repository.tabulated_benchmark import BlackboxRepositoryBackend
 from benchmarking.nursery.benchmark_kdd.baselines import MethodArguments, methods
 from benchmarking.nursery.benchmark_kdd.benchmark_definitions import benchmark_definitions
 
 from syne_tune.backend.simulator_backend.simulator_callback import SimulatorCallback
-from syne_tune.optimizer.schedulers.transfer_learning import TransferLearningTaskEvaluations
 from syne_tune.stopping_criterion import StoppingCriterion
 from syne_tune.tuner import Tuner
 from coolname import generate_slug
